+++ tests/unit/test_translator.py
@@ -63,8 +63,8 @@ class ModelPbToEntityPbTranslatorTestCase(unittest.TestCase):
 
         entity_pb_translated = model_pb_to_entity_pb(model_pb=example_pb)
 
-        self.assertEqual(entity_pb_native, entity_pb_translated)
         self.assertEqual(repr(entity_pb_native), repr(entity_pb_translated))
+        self.assertEqual(entity_pb_native, entity_pb_translated)
         self.assertEqual(sorted(entity_pb_native.SerializePartialToString()),
             sorted(entity_pb_translated.SerializePartialToString()))
 
@@ -101,8 +101,8 @@ class ModelPbToEntityPbTranslatorTestCase(unittest.TestCase):
 
         entity_pb_native = datastore.helpers.entity_to_protobuf(entity)
 
-        self.assertEqual(entity_pb_native, entity_pb_translated)
         self.assertEqual(repr(entity_pb_native), repr(entity_pb_translated))
+        self.assertEqual(entity_pb_native, entity_pb_translated)
         self.assertEqual(sorted(entity_pb_native.SerializePartialToString()),
             sorted(entity_pb_translated.SerializePartialToString()))
 
@@ -413,8 +413,8 @@ class ModelPbToEntityPbTranslatorTestCase(unittest.TestCase):
         # native one)
         entity_pb_native = datastore.helpers.entity_to_protobuf(entity)
 
-        self.assertEqual(entity_pb_translated, entity_pb_native)
         self.assertEqual(repr(entity_pb_native), repr(entity_pb_translated))
+        self.assertEqual(entity_pb_translated, entity_pb_native)
         self.assertEqual(sorted(entity_pb_native.SerializePartialToString()),
             sorted(entity_pb_translated.SerializePartialToString()))
 
@@ -424,6 +424,42 @@ class ModelPbToEntityPbTranslatorTestCase(unittest.TestCase):
         self.assertEqual(example_pb_converted, example_pb)
         self.assertEqual(sorted(example_pb_converted.SerializePartialToString()),
             sorted(example_pb.SerializePartialToString()))
+
+    def test_model_pb_to_entity_pb_repeated_referenced_field_with_enum_field(self):
+        # Test a scenario where a repeated field references a nested type which contains an ENUM
+        # and ensure that default enum value (0) is correctly set either when it's explicitly
+        # provided or when it's not provided and a default value is used.
+        example_pb = example_pb2.ExampleDBModel()
+
+        example_placeholder_pb1 = example_pb2.ExampleNestedModel(string_key=u'value 1',
+            int32_key=12345, enum_key=example_pb2.ExampleEnumModel.ENUM2)
+        # Enum with value 0 is explicitly provided
+        example_placeholder_pb2 = example_pb2.ExampleNestedModel(string_key=u'value 2',
+           int32_key=5000, enum_key=example_pb2.ExampleEnumModel.ENUM0)
+        # Enum value is not provided, default value 0 should be used
+        example_placeholder_pb3 = example_pb2.ExampleNestedModel(string_key=u'value 3',
+            int32_key=40)
+
+        example_pb.complex_array_key.append(example_placeholder_pb1)
+        example_pb.complex_array_key.append(example_placeholder_pb2)
+        example_pb.complex_array_key.append(example_placeholder_pb3)
+
+        self.assertEqual(example_pb.complex_array_key[0].enum_key, 2)
+        self.assertEqual(example_pb.complex_array_key[1].enum_key, 0)
+        self.assertEqual(example_pb.complex_array_key[2].enum_key, 0)
+
+        # Serialize it and ensure "0" enum values are included
+        entity_pb = model_pb_to_entity_pb(model_pb=example_pb)
+        self.assertEqual(len(entity_pb.properties['complex_array_key'].array_value.values), 3)
+        self.assertEqual(entity_pb.properties['complex_array_key'].array_value.values[0]
+                .entity_value.properties['enum_key'].integer_value,
+            example_pb2.ExampleEnumModel.ENUM2)
+        self.assertEqual(entity_pb.properties['complex_array_key'].array_value.values[1]
+                .entity_value.properties['enum_key'].integer_value,
+            example_pb2.ExampleEnumModel.ENUM0)
+        self.assertEqual(entity_pb.properties['complex_array_key'].array_value.values[2]
+                .entity_value.properties['enum_key'].integer_value,
+            example_pb2.ExampleEnumModel.ENUM0)
 
     def assertEntityPbHasPopulatedField(self, entity_pb, field_name):
         # type: (entity_pb2.Entity, str) -> None